Start implementing code generation
I'm able to compile and run test.scm, which successfully prints 3 to
stdout. The standard library is nowhere near complete, and in
particular, I want to be able to work with values other than integers.

I brought in petewarden's simple hashmap implementation for C because
I'm not interested in implementing a hashmap for the purposes of this
project.