Name normalization for multi-module support
Multi-module support introduced two problems with naming:

- Modules could be named with characters that are invalid as assembler
  labels (such as using a dash). Multi-module support added the module
  name as part each label.

- Members of modules could be named with such characters as well. This
  was not initially a problem because all symbols were looked up with
  the name as a string, specified as the key into a hash table. Now,
  some of these symbols are exported, and they generate globally
  accessible assembler labels.

One option is to simply strip out (or replace) these offending
characters. However, that would cause a problem if we had two modules or
members that differed only by that character (e.g. a-b and ab). Thus,
the solution is to do two things:

1. Replace these characters.
2. Append a hash of the *original* name.