Fix varargs support
Now, user-defined functions can be variadic, and the functionality
works. It turns out the stack alignment logic was quite wrong in many
cases, so I re-worked it heavily to fix it. This allowed varargs to work
correctly.

I haven't tested this with lambdas, but it should work. The caveat is
when we want no positional arguments in a lambda, which requires
slightly modified syntax that will be added later.