Start of failure tests
A key observation is that any test may fail during either the
compilation or the running stage.

In the case of the success tests, I added some granularity so users can
see which stage failed. For failure tests, the condition we want to
avoid is if *both* run successfully, so that granularity is not needed.

I also added some helpful logging, including outputting the currently
running test, which is then cleared after the test finishes running.