Separate parsing and code generation
This will be useful for future static analysis, but for now, I will use
the intermediate structure to store the type of module that is being
compiled (scheme vs. C for example).

While this was not strictly necessary for C module support, it's better
to do this refactoring now than to do it later when this type of logic
is more intertwined and complicated.