Consolidate module getters into one function
Instead of having a separate getter functions, one for each exported
symbol, we can use a single function that takes in the name of the
symbol to get as a string.

The disadvantage currently is that we're no longer preventing access to
non-exported symbols. However, this can be solved either through static
analysis (since the information is present), or even by storing those
objects in a different frame used only by the getter.

The *advantage*, however, is that we can generate a proper wrapper
around native modules, where we *don't* know what all the exported
symbols are.