First step in C module support
Now, it's actually possible to call native functions. However, much is
left to be done. In fact, the first thing to do next is to determine how
to pass arguments, since as far as the caller is concerned, a scheme
function is being called, and so the argument should be passed via the
stack, not via the registers.

Supporting arguments *could* result in some major changes, so this
commit is just the start in a longer journey.