First step in arguments passing to C functions
I've just made function calls more expensive :(

Now, every time a function is called, either scm_fncall is called or
native_fncall is called based on checking a field inside the lambda
structure.

In order to get this infrastructure up and running, I passed the
arguments as a cons list to the C function. I'm not happy with this, and
I'm going to see if I can pass along individual arguments.