Remove some overhead of scm function calls
This is basically a tail call elimination. Since calling the scheme
lambda is the last step of scm_fncall, it makes sense for it to not put
its own address on the stack and let the called function return directly
to the original caller.

The only caveat is that the callee has to know to remove the stack frame
that was added by scm_fncall from the stack before returning. This does,
however, have the advantage of not duplicating the stack frame before
and after the return address to scm_fncall.