New syntax for calling native functions
While I had worked hard to avoid this, this is a way better developer
experience than having to deal with cons lists in C functions, and much
faster to boot (no more checking if a function is native or not at
runtime).

I can iterate on this in the future, but this is a suitable compromise.