Add SDL library to standard lib
First use of C-interoperability! The next step is figuring out a nice
way to pass the correct arguments to the compiler so that libraries such
as SDL can be included in the compilation process in a cross-platform
way (right now, I'm always including the arguments, and that too
specifically for OS X).