Start on basic C code parsing
I experimented with using cast to parse C files, but it requires the C
preprocessor to run in order to resolve typenames (and rightly so).
However, in the common case, it should be possible to extract the list
of exports. This won't work if the list is constructed using CPP magic,
but again, this is the common case we're dealing with.

This pseudo-parser requires that the arity of the C functions are
present in the export list, which we'll need at compile time. However, I
haven't changed any of the C files to reflect this yet.