Finally unify native and non-native functions
By doing some static analysis of the export list in C files, we can
create a wrapper around the native functions that pull from the stack
and populate the argument registers.

Finally, it's now possible to call a native function as if it's just any
other function written in the language!

The ccall syntax actually doesn't work anymore, but I'll get rid of it
in a subsequent commit.