Tests: separate compile and runtime failures
Ideally, most issues would be caught by caught by the compiler, but
there are cases where this does not happen:

- Unimplemented features in the compiler.
- Cases where the language doesn't provide enough information for the
  compiler to know that the result is incorrect.

It's good to create tests as early as possible, even if it means the
compiler won't catch errors in order to ensure the program fails at run
time. Later, these tests can be moved to the compile-time failure bucket
once the necessary features are implemented.