Proper let/let*/letrec semantics
This is achieved via static analysis of the names that are in scope as
the bindings are evaluated.

With this change, all the runtime failure cases have been made into
compile-time failure cases, and I've removed the runtime failure test
category altogether. It can be added in when there's a legitimate need,
but otherwise, I'm hoping to make all other failure cases fail at
compile time.