Add default "else" class to all cond expressions
This is so that if a "cond" expression has no matching conditions and no
"else" clause, it still returns "nil". Other implementations seem to
have some sort of "undefined" or "void" value that's returned (which is
also what's returned by functions such as "display"), but we don't have
that so we'll make do with "nil" for this case.