Scheme -> Garlic everywhere
I still need to clean up the C-related files, such as scm.h. I might
keep the .scm extensions though, if for nothing else but to provide
default syntax highlighting in editors or code highlighters.

I changed the Github repository name, so the README.md instructions are
valid as of now.