Move a few "stdlib.c" functions to "core.c"
"core.c" is a new C module that will always be included in all garlic
modules. The reason for this is to cut the fat from what is actually the
runtime (as opposed the the standard library).

I kept some functionality inside the runtime file because it's
functionality I expect will be used by C modules as well, and so they're
in a sense part of the runtime. I probably want to move these functions
into a separate "support" library, and rename "stdlib.c" to "runtime.c".

As a result of this change, "stdlib.S" is leaner too.

I'll have to implement variadic C functions in order to move the
remaining stdlib functions into "core.c".