Leave extra space on stack when calling functions
This extra space is for a single scenario: the function being called is
a variadic C function (not yet implemented), and no extra arguments are
passed to the function.

When this happens, the gathered extra arguments is simply "nil", but it
needs to be added to the "bottom" of the stack. This means we need to
leave an extra space for that argument.

The reason this isn't needed for variadic garlic functions is because
when a garlic function is invoked and it gathers up extra arguments, the
resulting list is added directly to the current frame. There is no such
frame for C functions (or rather, the stack *is* the frame), hence the
need for extra space on the stack.