Move "display" to core.c
This is the last function that was in stdlib.c and that was added to the
root frame by stdlib. This means that we no longer need to have any
concept of "global symbols" because these symbols will always be seen by
the module system since they come from an imported C module.