Start of recursive compiler
I'm focusing first on the parser, in which I assume I have a series of
tokens coming from the lexer (currently mocked out to return test
streams) and I construct an AST.

Instead of doing a codegen right away, I'm going to practice doing AST
transformations by implementing some semantic analysis and pretty
printing.

In this process, I've already found parts of the standard
library/language that are missing, and I'll continue to find more. At
the end of this process, I'm hoping the language will be that much more
useful.