Use test "evaluator" to test that parser works
I don't want to build out a full interpreter, as that would be throwaway
work which would need to be re-implemented when I write a proper
back-end for the recursive compiler. However, having a simple evaluator
means I can test that the AST representation is usable and the parser
behaves correctly.