Use Global Offset Table on Linux
I'm not sure if a recent GCC update or some other environmental factor
was the culprit, but now, I need to explicitly use the Global Offset
Table (GOT) for Position Independent data accesses. Luckily, I alread
had to do that on OS X, so now I'm just going to do the same on Linux as
well.