Include original filename in generated assembly
This is a change I should I have made a long time ago: in the generated
assembly files, include the original filename so it's easy to see which
assembly file corresponds to which source file!

The filename is printed relative to the current working directory. I
don't think this is ideal, because now the generated output depends on
where the compilation happens, but I find it better (for both debugging
and security) than the alternatives:

- Print just the filename, no path information. I would rather be able
  to disambiguate when two modules with the same name are present in
  different directories.

- Print the absolute filename. I don't want to leak information about
  the entire computer into the generated output.