[recursive] return errors from lexer
This commit is a big revamp of the lexer to support returning actual
errors. Previously, as soon as an error was detected, the compiler would
print the error and exit. Now, the lexer tracks the error, still
choosing to stop lexing at the first error, but returning the error
instead of panicking. This has a few advantages over the previous
approach:

1. Lays some groundwork for how I plan to report errors from the parser.

2. This paves the way for reporting multiple errors instead of stopping
   at the first one. I do have a few ideas of where I would use this in
   the lexer, but the real benefit will be in the parser.

3. Packaging up multiple pieces of data into a unified "state", as was
   necessary to pass around both results/errors and the remaining
   unlexed strings lays the groundwork for passing around more data
   throughout the lexing process. Where I plan to use this is for
   reporting the _location_ of the error, something even the current
   Ruby implementation struggles with.

I've generally tried to focus on the refactoring in this commit, instead
of introducing new functionality. In particular, it was always possible
to detect unterminated strings even before this change. But with this
framework in place, I hope to add in a few other checks as well, like
invalid integers.

Even though these errors were being reported even with the previous
recursive implementation, I've explicitly added these syntax errors as
test cases for posterity.