[recursive] report locations of lexer errors
As promised in the previous commit, the refactoring to pass around
multiple pieces of data during the lexing process is what made it easy
to incorporate one more piece of data: the location in the source code
where each part of the lexing happens.

This in turn allows that location to be included in any errors, making
the error reports much clearer. So, even though the recursive compiler
is way behind the Ruby implementation, in this one aspect, the recursive
compiler is already ahead.