[recursive] reject non-delimeters immediately after integer
This is a new error detection that was not present in the existing the
lexer prior to the refactor to handle the error reporting. While this
functionality could have been added earlier, adding it now demonstrates
how the error handling is able to report this error, and how easy it is
to incorporate said error handling into the new error reporting
structure.