[recursive] include source locations in lexed tokens
This has no user-facing effect as of yet, but it will be useful when
reporting parser errors in the future.

Because tokens are accessed using strong encapsulation, pretty much no
logic that consumes tokens in the parser had to be modified. The only
changes were to the lexer constructing new tokens and the parser
synthesizing tokens to convert between equivalent forms. The latter may
prove to be a bit of a problem when reporting errors, but that will have
to be addressed when adding error reporting to the parser.