[recursive] report errors during tree generation
Reports errors related to mismatched parentheses.

This is the first step in adding error reporting to the parser phase of
the compiler. The most notable part that's missing is error reporting
during generation of the AST, when the tree structure is actually imbued
with meaning and the tokens converted into AST nodes.

Also added some test files that exercise the newly handled cases. Once
again, the recursive compiler reports the errors better than the Ruby
implementation!