Add "set" data structure to standard library
This data structure is by no means efficient, but it gets the job done.

The data structure will be used for the static analyzer of the recursive
compiler, to track what names are in scope. In this usage, the set of
names will be generally small enough for the inefficiency to not matter.
At the same time, the data structure is not really specific to the
recursive compiler, so I'm including it in the standard library for any
program to use.