[recursive] move program vs. section headers in hand-crafted ELF
The main reason for doing this is because, in my own visualization
(still to come on my blog or somewhere else), I'm showing bytes in runs
of 16 bytes. The section headers are 0x40 bytes long, which is a
multiple of 16, but the program headers are 0x38 bytes long, which is
not a multiple of 16.

Putting the section headers first means the file is "nicely aligned" for
longer, before getting to the "less aligned" portions.

The changes are:

- Physically move the bytes.
- Update the start offsets for both the program and section headers
  tables in the ELF header.