Allow comments in a few more places
The actual bug I encountered when writing the recursive compiler was I
couldn't add a comment without a leading space before the semicolon in
certain cases.

This happens because _inside_ a list, the Ruby-based Parslet grammar I
specify requires spaces separating expressions with spaces. In most
cases, this is desired, and in fact, this should also be required in
certain cases at the top level. For example, nowhere should the
expressions `1"str"` be allowed as two sequential expressions.
Unfortunately, this assumption is wrong for comments.

Fixing this bigger issue is proving to be difficult with the Parslet
grammar. I'm not sure the hand-written parser in the recursive compiler
will fare any better.

However, I did add some new test cases that the recursive compiler
should eventually pass, when I get to that point. In doing so, I
extended the non-recursive compiler to allow comments in a few places it
did not allow in the past.