Convert values to string during display, regular or error
The `display` function was previously implemented by simply printing out
strings during the "value -> string" conversion process. This allowed
handling different types of values, but the stringification process was
not reusable in any context that didn't involved just printing the
strings directly.

In particular, the conversion to string was not possible for the
`error-and-exit` function, which had previously assumed all the
arguments to the function were already strings. As a result, I got
unexpected errors when implementing the recursive compiler when I tried
to exit with an error involving a non-string value.

In this commit, I first convert all the values to either the `display`
or the `error-and-exit` function into one big string. Then, I either
print that string or send it over to the runtime's error and exit
mechanism. Note: there are plenty of memory leaks, but that's the case
for pretty much the entire compiler :(