[recursive] resolve text section offset/memory addresses in program header
To achieve this, I first associated with each section stub an ID and
placed that ID in each section and program header stub. This way, when
processing the program headers, I can look up the original section stub
and check its offset to compute both the file offset and the memory
addresses.

With the basic infrastructure in place, I will now apply this same
principle to the section header table entries in subsequent commits.