[recursive] compute file offsets in sections headers
One refactoring here is to treat the section header string table (the
`shstrtab`) as a stub in and of itself, instead of treating it
specially. I was already creating a stub for its associated section
header, but not for the section itself.

With that change, I'm able to use the laid out stubs to compute the
offsets as needed by the section headers table.