[recursive] resolve memory addresses in section headers
Unlike the file offsets, the memory addresses pre-layout may not be
placeholders. This is important for sections that don't get loaded into
memory, like the section header string table.

Technically, for those sections, constructing a memory address should be
okay, as the address field in the section header should be ignored for
that section. However, I wanted to make sure the address is explicitly
set to all zeros for clarity.