[recursive] calculate entrypoint in ELF header
I decided that instead of a new field to denote a specific text stub as
the entrypoint, I would just use a specific ID for that stub.

At this point, the entire ELF generation works for this simple case!
What's left is to actually write the bytes to a file so I can execute
the file and ensure it works. That said, given the output bytes are
identical to my test file, the output should work.