[recursive] compile simple programs containing only int statements
This is the barebones "compilation" functionality. While almost nothing
is supported, the compiler now has the ability to take an int statement
and generated code to output that integer into `%rax`. The final result
is then returned as the status code of the executable.

All of this shows that the lexer + parser + static analyzer + code
generation + elf file generation work together in tandem to implement a
full compilation pipeline!