[recursive] codegen conditionals
Because the parser already converts `if` into `cond`, both types of
conditionals are automatically supported, as long as all sub-expressions
(conditions and bodies for the clauses) are supported by the codegen.

Due to the copious error handling, if any problems arise with the
codegen--such as a sub-expression not being supported by the codegen--I
continue to get clear error messages.