Support two's complement for int -> little endian conversion
This doesn't practically affect the ELF generation, as all the numbers
that are converted using this utility are unsigned integers (usually
addresses and sizes). This is verified by the automated ELF generation
test.

 However, as I make progress on code generation, I will need to support
signed/negative values.

To ensure the conversion code is correct and continues to be correct, I
added an automated test. This isn't a test of any standard library or
core functionality, but I also don't have a good place to put tests that
are specific to one project.