[recursive] support adding a .data segment into ELF file
The same layout and offset resolution algorithm I already developed for
the overall ELF file generation applies here. The main difference is
that the data section is loaded into a later part of the virtual address
space compared to text sections.

What's missing still is the ability to add a placeholder in the `.text`
section that will then be resolved into the offset from the data section
based on where that section ends up in the final file.