Add janky string->bytes conversion
The encoding story is a mess. Ideally, I should have built the string
type in Garlic with a specific encoding in mind (and even better, allow
for different encodings!). But for now, I'm using `char *` under the
hood, and I'm mostly dealing with ASCII.

Luckily, by ensuring strings in Garlic are wrapped in a custom data
type, the internal representation can be switched out. C extensions will
need to update their code when unwrapping strings, but I'm not too
worried.