[recursive] minor fix for adding .data section to ELF
When computing offsets for each entry in the data section, I was first
calling `assoc:pairs` to get the unique entries. The order of the
entries returned by this helper is undefined, and in practice, it ended
up being the reverse order as the original.

However, when putting all the bytes together, I was operating on the
original entries, before calling `assoc:pairs`. This had two
consequences:

1. I might have accidentally included duplicate entries in the final
   byte contents for the section. This was the entire point of calling
   `assoc:pairs` when computing the offsets.

2. The order in which the offsets were computed ended up different than
   the order in which the bytes were collated, even if there were no
   duplicates.

To make the two operations consistent, I called `assoc:pairs` once on
the original input and performed the two operations on the de-duplicated
list of pairs, whatever the order of that list is.